+++ tests/fixtures/actions.py
@@ -1,3 +1,6 @@
+import os
+from copy import deepcopy
+
 EmptyAction = {}
 
 
@@ -6,7 +9,6 @@ DoNothingAction = {
     "type": "action",
     "provider": {"type": "python", "module": "fixtures.fakeext", "func": "do_nothing"},
 }
-
 
 EchoAction = {
     "name": "a name",
@@ -40,3 +42,22 @@ InterruptAction = {
         "func": "force_interrupting_experiment",
     },
 }
+
+PythonModuleActionWithLongPause = {
+    "type": "action",
+    "name": "action-with-long-pause",
+    "pauses": {"before": 30, "after": 5},
+    "provider": {
+        "type": "python",
+        "module": "os.path",
+        "func": "exists",
+        "arguments": {
+            "path": os.path.abspath(__file__),
+        },
+        "timeout": 40,
+    },
+}
+
+PythonModuleActionWithLongAction = deepcopy(PythonModuleActionWithLongPause)
+PythonModuleActionWithLongAction["pauses"]["after"] = 30
+PythonModuleActionWithLongAction["pauses"]["before"] = 35